feat(#24): page 컴포넌트로 분리
