+++ src/pages/game/FinalDebate.jsx
@@ -1,17 +1,17 @@
-import React, { useState, useRef } from 'react';
+import { useState, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import useAutoScroll from '../../hooks/useAutoScroll';
 
 import ChatInput from '../../components/chat/ChatInput';
-import Timer from '../../components/Timer';
+import Timer from '../../components/game/Timer';
 
-import char1 from '../assets/images/char1.png';
-import char2 from '../assets/images/char2.png';
-import char3 from '../assets/images/char3.png';
-import char4 from '../assets/images/char4.png';
+// import char1 from '../assets/images/char1.png';
+// import char2 from '../assets/images/char2.png';
+// import char3 from '../assets/images/char3.png';
+// import char4 from '../assets/images/char4.png';
 
-const FinalDebatePage = () => {
+const FinalDebate = () => {
   // 채팅 메세지 더미
   const [messages, setMessages] = useState([
     { senderId: 1, name: '산책하는 노루', avatar: char1, message: '억울해요' },
@@ -128,7 +128,7 @@ const FinalDebatePage = () => {
   );
 };
 
-export default FinalDebatePage;
+export default FinalDebate;
 
 const PageContainer = styled.div`
   display: flex;
