+++ src/store/step.js
@@ -1,0 +1,7 @@
+import { create } from 'zustand';
+
+export const useStepStore = create((set) => ({
+  step: 0,
+  nextStep: () => set((prevState) => ({ step: prevState.step + 1 })),
+  prevStep: () => set((prevState) => ({ step: prevState.step - 1 })),
+}));
